requirements.txt
python-telegram-bot==5.3.0
requests==2.13.0
twilio==5.7.0
